Add Stack Navigator as root
