+++ .gitignore
@@ -13,3 +13,6 @@ web-build/
 # macOS
 .DS_Store
 package-lock.json
+
+# Visual Studio Code
+.vscode
