+++ screens/MyMentors.js
@@ -1,13 +1,25 @@
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import React from 'react';
 import { Header } from '../components/Header';
-import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
+import { Button } from "@rneui/themed";
+import { Text } from '@rneui/themed';
+
 
 export default function MyMentors() {
   return (
     <>
       <Header view="Home" title="My mentors" />  
-      <Text>MyMentors</Text>
+      {/* <StatusBar style="auto" /> */}
+      <View style={styles.container}>
+        <Text h4={true}>You don't have any mentor yet!</Text>
+        <Button
+          style={styles.findMentorsButton}
+          title="Find Mentors"
+          onPress={() => {
+            navigation.navigate("MyMentors");
+          }}
+        />
+      </View>
     </>
   )
 }
@@ -19,4 +31,7 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
   },
+  findMentorsButton: {
+    marginVertical: 10
+  }
 });
